move all task/curriculum loading logic away from their classes, into a new submodule
