+++ test/curriculum/test_learning_task.py
@@ -131,89 +131,6 @@ class TestLearningTask(unittest.TestCase):
         with self.assertRaises(ValueError, msg='Unknown stop condition should raise an error'):
             _get_learning_task(mock_env, stop_conditions={'unknown_stop': 123})
 
-    def test_loading_config(self, mock_env: ScalableEnvironment, mock_agent: Agent):
-        """
-        ``LearningTask`` should be able to load a configuration from a YAML file
-        """
-        with tempfile.TemporaryDirectory() as temp_dir:
-            task_config = ("env_type: placeholder\n"
-                           "env_args:\n"
-                           "  difficulty: 0\n"
-                           "stop_conditions:\n"
-                           "  max_episodes: 2\n"
-                           "evaluation_interval: 24\n"
-                           "name: Reece James\n"
-                           f"agent_save_path: {temp_dir}/secret_agent_123\n"
-                           f"stats_save_path: {temp_dir}/super_stats_321")
-            config_file_path = os.path.join(temp_dir, 'config.task.yml')
-            with open(config_file_path, 'w') as f:
-                f.write(task_config)
-            # load config - it should be identical to the task defined above
-            # patch to avoid error when loading the mock environment
-            with mock.patch.object(LearningTask, 'get_type',
-                                   mock.MagicMock(return_value=lambda *args, **kwargs: mock_env)):
-                sut = LearningTask.load(config_file_path)
-
-        # run the task to be able to check some of the parameters validity later
-        sut.run(mock_agent)
-
-        self.assertEqual('Reece James', sut.name)
-        self.assertEqual(24, sut.stats.evaluation_interval)
-        self.assertEqual(f'{temp_dir}/secret_agent_123', sut.agent_save_path)
-        self.assertEqual(f'{temp_dir}/super_stats_321', sut.stats_save_path)
-        # stop condition
-        self.assertEqual(2, len(sut.stats.episode_rewards))
-
-    def test_saving_loading_config(self, mock_env: ScalableEnvironment, mock_agent: Agent):
-        """
-        Task's configuration should be saved in a way that loading it will produce a task of
-        identical configuration
-        """
-        task_to_save_config = {
-            'name': 'Frank Lampard',
-            'evaluation_interval': 8,
-            'include_init_eval': False,
-            'evaluation_count': 26,
-            'greedy_evaluation': False,
-            'exploration_reset_value': 33,
-            'agent_save_path': './agent_path',
-            'stats_save_path': './stats_path',
-        }
-        task_to_save = _get_learning_task(mock_env, other_task_args=task_to_save_config)
-        with tempfile.TemporaryDirectory() as temp_dir:
-            save_path = task_to_save.save(os.path.join(temp_dir, 'test.task.yml'))
-            self.assertTrue(os.path.exists(save_path))
-            # load config - it should be identical to the task defined above
-            # patch to avoid error when loading the mock environment
-            with mock.patch.object(LearningTask, 'get_type',
-                                   mock.MagicMock(return_value=lambda *args, **kwargs: mock_env)):
-                sut = LearningTask.load(save_path)
-        sut_config = sut.to_dict()
-        sut_config.pop('env_type')
-        sut_config.pop('env_args')
-        sut_config.pop('stop_conditions')
-        self.assertEqual(task_to_save_config, sut_config)
-
-    def test_saving_loading_path(self, mock_env: ScalableEnvironment, mock_agent: Agent):
-        """
-        Saving and loading using the same path should always work, regardless whether an expected
-        extension is provided or not.
-        """
-        sut_save = _get_learning_task(mock_env)
-        with tempfile.TemporaryDirectory() as temp_dir:
-            path_no_extension = os.path.join(temp_dir, 'config')
-            path_extension = os.path.join(temp_dir, 'config.task.yml')
-            sut_save.save(path_no_extension)
-            sut_save.save(path_extension)
-            # patch to avoid error when loading the mock environment
-            with mock.patch.object(LearningTask, 'get_type',
-                                   mock.MagicMock(return_value=lambda *args, **kwargs: mock_env)):
-                try:
-                    LearningTask.load(path_no_extension)
-                    LearningTask.load(path_extension)
-                except FileNotFoundError:
-                    self.fail('save() and load() path resolving should match')
-
     def test_include_init_eval_param(self, mock_env: ScalableEnvironment, mock_agent: Agent):
         # arrange
         sut_init_eval = _get_learning_task(
